Rename 'variables' to 'conditionals'
File deals with all conditional executions
