+++ greta/__init__.py
@@ -1,4 +1,4 @@
 __version__ = "0.1.0"
 
 from .checks import check_intensity
-from .variables import condition_variable
+from .conditionals import condition_variable
